another changes

+++ app/list_delete.php
@@ -1,14 +1,15 @@
 <?php
 
-if ( $_SERVER[ "REQUEST_METHOD" ] != "DELETE" ) {
-    die( $_SERVER[ "REQUEST_METHOD" ] );
+session_start();
+// DATA FORM REQUEST
+$request_data = json_decode( file_get_contents( "php://input" ) );
+
+if ( $_SERVER[ "REQUEST_METHOD" ] != "DELETE" || $_SESSION['ID'] != $request_data->id_user) {
+    die( include "../404.php" );
 }
 
 require_once "../_utilities/Task_List.php";
 $list = new Task_List();
-
-
-$request_data = json_decode( file_get_contents( "php://input" ) );
 
 $response_data = $list->delete_list( $request_data );
 
@@ -17,3 +18,5 @@ header( "Access-Control-Allow-Methods: DELETE" );
 header( "Content-Type: application/json; charset=UTF-8" );
 
 echo json_encode( $response_data );
+
+die();
